+++ src/encoded/types/individual.py
@@ -17,30 +17,30 @@ def _build_individual_embedded_list() -> List[str]:
         "father.accession",
         "father.is_deceased",
         "father.sex",
-
         # Individual linkTo
         "mother.accession",
         "mother.is_deceased",
-        "mother.sex"
+        "mother.sex",
     ]
 
 
 @collection(
-    name='individuals',
-    unique_key='accession',
+    name="individuals",
+    unique_key="accession",
     properties={
-        'title': 'Individuals',
-        'description': 'Listing of Individuals',
-    })
+        "title": "Individuals",
+        "description": "Listing of Individuals",
+    },
+)
 class Individual(Item):
-    item_type = 'individual'
-    name_key = 'accession'
-    schema = load_schema('encoded:schemas/individual.json')
+    item_type = "individual"
+    name_key = "accession"
+    schema = load_schema("encoded:schemas/individual.json")
     rev = {
-        'children_f': ('Individual', 'father'),
-        'children_m': ('Individual', 'mother'),
-        'families': ('Family', 'members'),
-        'case': ('Case', 'individual')
+        "children_f": ("Individual", "father"),
+        "children_m": ("Individual", "mother"),
+        "families": ("Family", "members"),
+        "case": ("Case", "individual"),
     }
     embedded_list = _build_individual_embedded_list()
 
@@ -48,47 +48,42 @@ class Individual(Item):
     def display_title(self, accession: str, individual_id: Optional[str] = None) -> str:
         return individual_id or accession
 
-    @calculated_property(schema={
-        "title": "Children",
-        "description": "Children of the individual",
-        "type": "array",
-        "items": {
-            "title": "Child",
-            "type": "string",
-            "linkTo": "Individual"
+    @calculated_property(
+        schema={
+            "title": "Children",
+            "description": "Children of the individual",
+            "type": "array",
+            "items": {"title": "Child", "type": "string", "linkTo": "Individual"},
         }
-    })
+    )
     def children(self, request):
-        kids = (self.rev_link_atids(request, "children_f") +
-                self.rev_link_atids(request, "children_m"))
+        kids = self.rev_link_atids(request, "children_f") + self.rev_link_atids(
+            request, "children_m"
+        )
         if kids:
             return kids
 
-    @calculated_property(schema={
-        "title": "Families",
-        "description": "Families this individual is a member of",
-        "type": "array",
-        "items": {
-            "title": "Family",
-            "type": "string",
-            "linkTo": "Family"
+    @calculated_property(
+        schema={
+            "title": "Families",
+            "description": "Families this individual is a member of",
+            "type": "array",
+            "items": {"title": "Family", "type": "string", "linkTo": "Family"},
         }
-    })
+    )
     def families(self, request):
         fams = self.rev_link_atids(request, "families")
         if fams:
             return fams
 
-    @calculated_property(schema={
-        "title": "Cases",
-        "description": "Cases for this individual",
-        "type": "array",
-        "items": {
-            "title": "Case",
-            "type": "string",
-            "linkTo": "Case"
+    @calculated_property(
+        schema={
+            "title": "Cases",
+            "description": "Cases for this individual",
+            "type": "array",
+            "items": {"title": "Case", "type": "string", "linkTo": "Case"},
         }
-    })
+    )
     def case(self, request):
         rs = self.rev_link_atids(request, "case")
         if rs:
@@ -103,10 +98,12 @@ class Individual(Item):
                 "title": "Primary Disorder",
                 "type": "string",
                 "linkTo": "Disorder",
-            }
+            },
         }
     )
-    def primary_disorders(self, disorders: Optional[List[JsonObject]] = None) -> Union[List[str], None]:
+    def primary_disorders(
+        self, disorders: Optional[List[JsonObject]] = None
+    ) -> Union[List[str], None]:
         if disorders:
             primary_disorders = set()
             for disorder_metadata in disorders: