+++ CHANGELOG.rst
@@ -6,10 +6,30 @@ cgap-portal
 Change Log
 ----------
 
-13.3.0
+13.4.0
 ======
+`PR 711: Generic QCs <https://github.com/dbmi-bgm/cgap-portal/pull/711>`_
 
 * Adds `QualityMetricGeneric` type
+
+
+13.3.1
+======
+`No PR: Staff-change Insert Updates`_
+
+* Remove Victoria's user insert from master-inserts
+* Add Cesar's user insert to master-inserts
+
+
+13.3.0
+======
+`PR 712: Somatic analysis updates <https://github.com/dbmi-bgm/cgap-portal/pull/712>`_
+
+* Add Individual linkTo on SomaticAnalysis schema
+* Add preservation method to Sample schema
+* Add primary disorders calcprop for Individual item
+* Add Individual embeds to SomaticAnalysis item
+* Update inserts for changes above
 
 
 13.2.1